requirements.txt
flask==3.0.3
python-dotenv==1.0.1
twilio==9.2.3
faiss-cpu==1.8.0.post1
numpy==1.26.4
pydantic==2.8.2
openai==1.43.0